+++ packages/koenig-lexical/src/components/ui/cards/MarkdownCard/MarkdownImageUploader.jsx
@@ -1,0 +1,19 @@
+import React from 'react';
+
+export default function MarkdownImageUploader({onChange, inputRef, loading}) {
+    return (
+        <div>
+            {loading && 'loading'}
+            <form onChange={onChange}>
+                <input
+                    name="image"
+                    type='file'
+                    accept='image/*'
+                    ref={inputRef}
+                    multiple
+                    hidden
+                />
+            </form>
+        </div>
+    );
+}
